fix(monitoring): use lastNotNull for non exported records
Switches the non exported records to use the last not null value, instead of the mean. It's not possible to provide a single number to indicate the value of this metric over a time span. To fix this we should stick to using the latest value of the time span. This provides the most usable information when analysing a cluster.