feat(broker): ensure log is flushed before committing snapshot
Broker can recover from a snapshot only if the follow events corresponding to the state included in the snapshot is available in the log. When explicit raft flush is disable, it is possible that a snapshot is committed, but the follow up events are not written to the disk. As a result, after a crash restart this snapshot may not be usable. To prevent such a state, we ensure that the necessary events are flushed before we commit the snapshot even if explicit flush is disabled.

(cherry picked from commit 92a3d5b2595ec970ee4dad234b0f96023a1acfe0)