feat(journal): prepare next segment asynchronously
Whenever a new segment is needed for writing to the journal, we also
trigger an async task to create the next segment ahead of time.
When this segment is then needed, the file is already created and filled
by the segment allocator. This reduces peak write latency because the
writer thread only has to wait for task completion without doing any
expensive IO.

The next segment is created uninitialized, without a journal descriptor.
On first access of the next segment, we initialize it by writing the
correct descriptor. When an uninitialized segment exists on recovery,
it is detected as a corrupted log segment (because it does not have the
expected descriptor) and thus deleted.

The segment id used to be a monotonically increasing number. Since we
increase this number before creating the next segment, any log resets
may break this assumption of monotonically increasing ids. In effect,
the log may start of with log segments 1-3, and an uninitialized segment
with id 4. When the log is then reset before initializing 4, log
segments 2 and 3 are deleted while 4 is not indexed and thus not
affected by the reset. When continuing to write to the log, segment 4
will be used, resulting in a consistent state of the journal with
segments 1 and 4. This may be undesirable for troubleshooting but causes
no other issues.