feat(clients/go): add standalone decision evaluation to zbctl
Support triggering standalone decision evaluation through zbctl. For
this purpose, a new root command is added: evaluate. When triggering a
decision evaluation, users can pass the decision ID or KEY. zbctl will
determine if the passed argument is an ID or KEY. Variables can be explicitly passed as a JSON
formatted string, or implicitly, through a path to a JSON file that defines the
variables.