refactor(logstreams): observe batch size in bytes
Adds a new metric to observe the batch size in bytes, giving us an idea
of the rate at which we "write" data on the leader.

To ensure we don't needlessly compute the batch size over and over,
since the batch is immutable, it's refactored to be computed once on
creation.