feat(logstream): log entries can be marked as `skipProcessing`
This uses an unused, previously reserved, byte in the
`LogEntryDescriptor` for arbitrary boolean flags. The first use case is
to mark entries that should be skipped during processing. This is
useful for batch processing where some commands that are written are
already processed and should be skipped the next time they are read.