refactor(atomix): remove blocking usage of futures
Removes unnecessary usage of futures in the append pipeline. We were
previously creating futures, but they were always completed and
processed in a blocking fashion. Since the cost of exceptions is less
throwing them and more creating them, we weren't really saving anything
there. Plus, we still had to handle the exception if it were to bubble
up. Finally, in all error cases, we were stepping down as well.