refactor(atomix): remove composable future
Removes the composable future utility. The only real production use case
was with the log compactor, which was returning a synchronous future
that was also never really used in an asynchronous fashion. This was
refactored to be synchronous (since that's how it's really used).

After that, the utility was removed since it isn't really doing much. If
we have more cases for this BiConsumer, we should add utility method but
avoid the inheritance we used here.