refactor(atomix): remove abstract and null thread context
Removes the abstract thread context class (now empty) and the null
thread context (now unused). Also removes the block toggling in the
OrderedFuture. OrderedFuture is kept as it's unclear if we rely on the
FIFO execution (as opposed to the CompletableFuture's LIFO execution) of
callbacks. This would require some more testing, but gut feeling is we
really don't. This was probably for the old session client/service proxy
mechanisms.