fix: handle batch size exceeding
Handle batch size exceeding which might happen inside a command processing. Rolls the current transaction
back (and the processing result). Retries the processing until the last successful processed command, as
we know this fits into the batch. The failing follow up commands will be then tried in a separate batch.