feat(test): verify dmn decision events are written
When we process a resource delete command for a DMN resource we should make sure we write all the required DELETED events.

This includes the DELETED event for the decision requirements. The DELETED event for each Decision within the DRG and the DELETED event for the resource deletion to indicate that it has finished processing.