feat(engine): process a DMN resource deleted command
When we process a ResourceDeletion.DELETE command we must check if the passed resource is a DRG. If it is we must write the DELETED events for the decision requirements and the decisions itself.

The appliers of these events will be responsible for making sure we delete resource from the state.

We don't have access to the namespace of the DRG here. As a result I have made the namespace of the DecisionRequirementsRecord an optional field.