test(journal): always close journal in tests
To prevent interference of temp directory cleanup by async preparation
of the next segment we must always close the journal before finishing a
test.

While the Journal is an `AutoCloseable`, restructuring all tests to use
try-with-resources made the tests less readable. Instead we always close
manually and remove the resource warning suppression to at least get a
warning when we forget to do so.