test: reliably fill journal for disk space recovery testing
This test would occasionally fail because after resuming exporting,
the broker disk did not free up enough to resume processing.

To fix this, we now publish messages with a larger variable payload
and with TTL=0. The larger payload fills up the disk faster while using
TTL=0 ensures that it is only the journal that is using the majority of
the disk space and that we don't accumulate too much state. This should
help when resuming exporting because the journal can be compacted,
freeing up disk space,  while the state cannot shrink
(until the message expires).