feat(engine): delete decisions from state
Delete decisions from the state when a Decision.DELETED event is written.
Apart from just deleting the decision from the relevant ColumnFamilies, it also needs to update the latest decision in the state. This is used to decide which decision is evaluated for a given decisionId.

If the deleted decision is the latest version, we need to find the previous version and set this as the new latest. If no other version is available the decision is removed from the latest decisions ColumnFamily.

If the deleted decision is not the latest version we van ignore this ColumnFamily entirely.