feat(engine): delete DRGs from state
Delete DRGs from the state when a DecisionRequirements.DELETED event is written.
Apart from just deleting the DRG from the relevant ColumnFamilies, it also needs to update the latest decision in the state.

If the deleted DRG is the latest version, we need to find the previous version and set this as the new latest. If no other version is available the DRG is removed from the latest ColumnFamily.

If the deleted DRG is not the latest version we can ignore this ColumnFamily entirely.