docs(README): increase minor release support
Previously, we mistakenly described the support for our minor releases
as 12 months. However, this should have been 18 months.

This updates the description to correct this mistake and adjusts the
illustrations of the minor and alpha release lifecycles.

To clarify the longer periods, I've added a calendar and changed the
example releases to real and current releases.