fix: don't re-wrap runtime exceptions thrown inside a transaction
When a `RuntimeException` is thrown inside a transaction, we should
never re-wrap this exception. This change ensures that all
`RuntimeException`s are forwarded as is. This makes exception handling
in the `ProcessingStateMachine` easier because it doesn't have to
inspect `RuntimeException`s to handle `ExceededBatchRecordSizeException`
correctly.

(cherry picked from commit d0d5228c7fd090caaa7e0a17dd215b4de2505f86)