fix(raft): do not resend vote request if a candidate cannot win election in one round
This can prevent other eligible nodes to become leader because both are competing with each other. Random delay would prevent this in most cases. But the RandomizedRaftTest caught a case where two nodes were in an election loop for a long time. Although in practice, the chances of this happening is very rare, it is better to decrease the probability even further.

With this fix, the candidate transition back to follower if it cannot win election in a round. After transitioning to follower, it has to send poll request again to become candidate. This let other eligible candidates to start election.