feat(protocol): add CommandDistribution intent
Adds the intents that are used for generalized command distributions.

STARTED - Will be used to store the record in the state. This way we can find it again in case we need to do a retry

DISTRIBUTING - Stores a pending distribution

ACKNOWLEDGE - Command to send acknowledgement of the distribution back to the distributing partion.

ACKNOWLEDGED - Removes a pending distribution

FINISHED - Will be used to remove the record from the state.