feat(journal): make flush thread-safe
Enables flush to be used in a thread-safe way by changing its behavior.
Flushing is now done on the segment, and relies on the fact that forcing
an unmapped buffer will throw an error. Since this is a very uncommon
case for us, as we unmap:

- On closing. It's not so crucial to be very fast here, and with careful
  ordering of the close operations we can avoid issues.
- On deleting the segment. This is a write-locked operation, and
  flushing is a read-locked, making them mutually exclusive (while
  keeping low contention with other read-locked operations).

Further more, flushing an unmapped buffer will not result in a
segmentation fault, but just an `UncheckedIOException` to be thrown.
Since this is a very uncommon case, we can simply handle it and suppress
the error if we know the segment was previously deleted/closed. In such
cases, we don't care to flush anyway.