refactor(journal): extract flushing logic for test purposes
Extracts the flushing logic, including its state, to allow for easier
testing in isolation. To do this, a minimal flushing API is also
introduced for the `Segment` class, `FlushableSegment`.

Management of the `lastFlushedIndex` and the general flushing logic is
moved to `SegmentsFlusher`, so we can properly test it. Otherwise the
old writer was too black-box-y to properly test.