fix(ci): optionally connect to internal Nexus
As a cost-saving measure a change was made to always connect to our internal Nexus instead of Artifactory. The internal Nexus is not publicly available and thus requires secrets.
Since a forked PR does not have access to the repository secrets (dependabot excluded) the change breaks the CI for external contributors.

This change makes it optional to have these secrets available. If they are not available we will connect to Artifactory, otherwise we will use the internal Nexus. The vast majority of the traffic will still use the internal Nexus, keeping the cost-saving measurements relevant.