fix(ci): do not trim xml output before uploading artifacts
When tests are run concurrently, the logs are interleaved in output.txt. It is not useful to debug when a test fails. The output from each test were removed from xml because large xml had caused issues when parsing the results. In this commit, we moved the trimming of xml files just before partsing the results. So if a test fails, the full xml is still available in the uploaded artifact.