feat(engine): visit msg deadlines with startAt
This adds the ability to visit the messages with expired deadlines
starting the visiting at a specific key (called Index).

This utilizes the new column family method whileTrue starting at key.
Note that the startAt Index is actually a composite key of the deadline
and the message key. When null is used the behavior stays like before:
it starts at the first entry of the column family.

A test has been added to verify the new behavior. Note that the
whileTrue method provided by the zeebe-db is tested more thoroughly. So
we don't have to test all edge cases of it. We can trust that this
method works as expected.

The visit method now also informs the caller when the visiting stops
about how it was stopped (by the visitor or because it did not find any
more entries to visit). This will become useful in the following commit,
where we want to determine how to continue after visiting the expired
messages.