ci: always remove system-out before publishing test results
This duplicates the existing logic from the `collect-test-artifacts`
action.

The goal is to move the removal from `collect-test-artifacts` to
`publish-test-results`. Because the `publish-test-results` action
is run from the `main` branch, removal needs to be a two-step process
of first adding it to `publish-test-results` and then removing it from
`collect-test-artifacts` in a separate PR.