ci: keep test output in artifacts
Now that the `publish-test-results` action trims the test output,
the `collect-test-artifacts` action does not need to remove it.
Keeping them is useful because for tests that run concurrently, the
system-out in the test report xml file is the only source for logs.

The usual `*-output.txt` files contain logs from the concurrent
execution of all tests within one class which makes them unusable for
debugging.