fix(protocol): fix random generation of CommandDistributionRecords
The CommandDistributionRecord contains a ValueType and a UnifiedRecordValue. We must make sure during the random generation of this record that the UnifiedRecordValue matches the ValueType. This is achieved by creating a custom randomizer for this specific class.

A downside of this implementation is that the ProtocolFactory is now aware of this specific Record. When we make changes in the Record we must change them in this randomizer as well.