refactor(journal): always store correct last flushed index
Ensures that resetting the last flushed index on reset/deleteAfter
always stores the correct index after the operation.

Additionally fix a bug, when truncating the last flushed index should be
the given index, not the one before, since we do keep the entry at the
index as before, meaning we should detect if it gets corrupted.