feat(java-client): add DeleteResource interface
Adds the interfaces for the DeleteResourceCommand. Both the command and the response interface are empty.

The command interface is empty because the command only requires a resourceKey, which will be passed as a parameter to the method creating the implementation of this interface. It implements the FinalCommandStep for sending the command.

The response interface is empty as the deleting a resource will not return anything.