test(engine): coverage for user task due and follow-up dates
Add test coverage for the new dueDate and followUpDate properties on
user tasks. These tests cover user task activation. They validate that
the new properties are attached to the JobIntent.CREATED event.

The tests also assert that incidents are raised if the expressions for
dueDate and followUpDate can't be properly evaluated due to a missing
variable, invalid date format, or invalid expression result type.