test(engine): verify distribution can be stored/retrieved/removed
This commit adds tests for the DbDistributionState. It verifies we can store pending distributions, and distributions itself. It also verifies we can retrieve the stored information. Finally, it also verifies the stored information can be removed again.

One thing to note is that the removing of a distribution must be idempotent. It could be that a partition sends the command back to the distributing partition twice. In this case we must be able to handle this command idempotent.
This is not the case for the pending deployments.