feat(engine): use foreign keys in DbDistributionState
There are 2 ColumnFamilies in the DbDistributionState. One uses the distributionKey to store the ValueType and RecordValue of the command that is getting distributed. The other uses the distributionKey and partitionId to keep track of the pending deployments.

A distribution begins with storing the ValueType and RecordValue. After this it will store a pending distribution for each partition it distributed too.
When a partition is finished with the distribution it will acknowledge this and the pending distribution is deleted. Once all pending distributions are deleted the ValueType and RecordValue van also be removed.

This means that as long as we have a pending deployment, the ValueType and RecordValue need to remains stored. We can guarantee this by using a foreign key.