fix(engine): limit expire msg commands in result
This adds a virtual limit to the number of EXPIRE Message commands the
MessageTimeToLiveChecker adds to a result when executed, to prevent that
it clogs the log stream with too many of these commands.

It achieves this by keeping track of the current visitor index, and
reusing this index in subsequent executions. The index is used to start
the visiting at a specific key in the column family. This was added in
the previous commit.

This index must be reset once the iteration reaches the point that there
are no messages with an expired deadline in the column family any more.
The reset is necessary to make sure we don't skip any messages whose
deadline has expired, but was added during or after the iteration.

The visit method informs the checker (when the visiting stops) whether
it was stopped by the visitor or not. If it wasn't, then this means the
end was reached and the index can be reset.

The checker reschedules itself. Either immediately, when it wants to
continue where it left off in another run, or after the interval if it
reached the end. Note that we need to continue to schedule the task
async.

The limit and the interval are not yet configurable, because this
requires changes to the `broker` module, i.e., exposing the config to
the engine. Since this module is shared between ZPA and ZDP, I prefer to
do this in a separate pull request, to make that easier to review.

Note that this commit breaks userspace, because the engine no longer
favors message expirations over regular processing and instead provides
a more equal balance between the two. It's unlikely that users depend on
this behavior.

(cherry picked from commit b4ff81eba6804225de7852fc775ec9b2019d5535)