fix: skip first entry in subsequent runs
Now that we use the lastIndex to start the visiting at, we visit the
last visited expired message a second time in a subsequent run.

The problem is that the visit method can only use a startAt not
startAfter index, while we can only keep track of the last index we
visited.

As a solution, we can simply skip the first entry we visit when it has
the same index as the last visited entry. This works because the index
is unique (deadline + message key).

A downside to this, is that there is one entry too many read from the
column family on every run.

(cherry picked from commit 980099fe7af3f9303c9bee655871df30028ebe57)