feat(stream-platform): introduce GatewayStream API
Introduces a new platform API to allow record processors to stream data
out to Zeebe Gateway instances.

Additionally adds concrete type implementations for the upcoming job
streaming feature, and an optional accessor in the
`RecordProcessorContext`.

Introducing this API as a first step will allow both the ZPA and ZDP
team to work mostly independently with minimal friction, even if it is
likely to change in the future.