refactor(engine): remove unused constructor
I want to add a parameter to the constructor, but discovered there was
this unused one as well. To reduce the effort, I just deleted this one.

The unused constructor was keeping the typedRecordProcessorFactory from
being final. So, now it can be made final as well.

(cherry picked from commit 76e13ec600149d84557451a64229b9a31aec0c1c)