feat(broker/engine): add engine config
The EngineCfg will be used for users to configure the engine at a system
level. It does not allow to configure the different engines per
partition. All configurations will be applied to the engine of each
partition.

This requires changes in both the broker and the engine. The engine does
not have a dependency on the broker. So the EngineCfg needs to be mapped
to an EngineConfig that is specified by the workflow engine.

(cherry picked from commit 63dcfb9cc9e7ae3265f91d2b924b4c1f6ff1103b)

Conflicts:
	broker/src/main/java/io/camunda/zeebe/broker/system/partitions/impl/steps/StreamProcessorTransitionStep.java
	engine/src/test/java/io/camunda/zeebe/engine/util/TestStreams.java