feat(broker/engine): add engine config
The EngineCfg will be used for users to configure the engine at a system
level. It does not allow to configure the different engines per
partition. All configurations will be applied to the engine of each
partition.

This requires changes in both the broker and the engine. The engine does
not have a dependency on the broker. So the EngineCfg needs to be mapped
to an EngineConfig that is specified by the workflow engine.