ci: cache maven dependencies explicitly to improve reliability
Many of our CI jobs would enable maven caching, some with a modifier,
some not.
Looking at the history of our CI jobs, caching appears to be not very
effective. One explanation for this is that all jobs would race for the
same cache, overwriting or invalidating what other jobs already saved.
Another possibility is that we simply exceed the cache limit of 10GB too
fast and evict useful caches too frequently.

This change introduces a dedicated maven-cache workflow which runs on
main and stable branches only. For each OS, we use the
`dependency:go-offline` maven target and save that in a cache keyed by
OS and a hash of all POM files. The `setup-zeebe` action only restores
from this cache and never saves to the cache again.
This makes the caching behavior very easy to understand. All
dependencies that are found by `dependency:go-offline` in the main
branch are cached, everything else is not.
This also reduces the number of caches. Effectively we only care about
3 cache lines, one for main and two for the stable branches.
Regular CI jobs no longer have to update the maven cache which might
save a couple of seconds on each job and removes one source of failures.

A negative effect from this is that feature branches that add new
dependencies will always download the new dependency because it will not
be cached until the dependency exists on main.
Additionally, `dependency:go-offline` is not exact and can miss some
dependencies. If this becomes an issue we can extend our `maven-cache`
workflow to explicitly download dependencies that are not found by
`dependency:go-offline`.