refactor(engine): expression processor handles empty strings for dates
When evaluating DateTime expressions with the ExpressionProcessor::evaluateDateTimeExpression method, allow for graceful handling of empty strings or null values instead of returning a failure. We want to allow this in order to conform to the business-oriented aspect of FEEL, i.e. an expression may return null or an empty string without raising an error.