feat: add additional async API
ZPA needs additional async API in order to schedule task on a separate actor, instead on a fixed rate
it is necessary to schedule the task with an delay which might be dynamic. This is the reason why
runWithDelay is the better fit.

Related to open PR https://github.com/camunda/zeebe/pull/11856