feat(engine): create CommandDistributionBehavior
The CommandDistributionBehavior is responsible to distribute commands to other partitions in a generic way. At this time it implements a single `distributeCommand` method. This takes a generic record. Using side-effects this generic record gets distributed using the InterPartitionCommandSender.

In order to verify that a distribution succeeds we create a CommandDistributionRecord. By writing a STARTED and DISTRIBUTING event for this record we can update the state to keep track of pending records. This will later be used to redistribute commands in case they never got acknowledged.