fix(engine): only update timestamp in new run
While visiting the expired messages in small iterations, we need to keep
the timestamp (that is used to compare which message have expired) the
same. Otherwise, we may continue to find new expired messages while
ignoring newly added messages with an earlier deadline.

This can be done quite easily, by keeping track of the current timestamp
to use for the visiting. We only update it at the start of a new run,
and reuse it in between the small iterations.