refactor(engine): split DeploymentCreateProcessor into 2 parts
The DeploymentCreateProcessor will have different behavior depending
which partition handles the command. If it's partition 1 it will need
to distribute the deployment. It also has to create timer events.

If it's a different partition we only have to make sure we store the
resources in the state and open start event subscriptions (apart from
timers).

This is step 1 of migrating this processor to the generalised
distribution flow.