fix(engine): migrate DeploymentClient expectations
As distributing a command now results in different records these
expectations need to be migrated to the new flow. When there is only 1
partition it is simple. We just need to wait for the Deployment CREATED
event.

When there's multiple event it becomes more complex. We want to make
sure the distribution has been completed on all partitions. For this we
need to look at the Deployment CREATED event on all partitions. We also
want to make sure the ACKNOWLEDGE commands have been handled by the
distributing partition. We can do this by looking for the
CommandDistribution FINISHED event.