test(engine): make sure visit returns correct result
There is a bug in the visitMessagesWithDeadlineBeforeTimestamp, because
it wrongly informs the caller that the visitor commanded to stop the
visiting, while that is not the case. This can happen when there are
messages available that are not yet expired.

This test case will fail at this time.

(cherry picked from commit e64778c9385725daebf3bcd14346d344615baa3b)