refactor: inline method
We're going to need to change the stoppedByVisitor boolean on line 360
depending on the if condition (deadline <= timestamp). We'll need to
inline the method to make that possible.

Note that the names deadline and messageKey were changed to
deadlineEntry and messageKeyEntry respectively. This is because the
method was static, but when inline this is no longer the case. The
variables would otherwise hide fields with the same names.

(cherry picked from commit 9b338a8261150688109c735ce6f6a0a9226e120c)