fix: correctly inform when stopped by visitor
This method was not correctly informing whether it was stopped by the
visitor. If there were more entries buffered that were not yet expired
then it should not continue. In that case it would also say that it was
stopped by visitor, but that's incorrect.

This rectifies that behavior.

(cherry picked from commit b3c6f6bc015f1ebb91964543121f7f5ff30b00e9)