feat(protocol): keep track of EventApplier versions
This commit changes the map of EventAppliers. It is changed to a map of VersionedEventAppliers. With this map we can decide which version of an event applier we should use.

We should NEVER use an event applier with a later version than the version on which the event was published. This will result in state inconsistencies. Instead we must use the latest version available that doesn't exceed the broker version at the time the event was published.