refactor: add failure message in topology loop
It sometimes happens that starter/worker are running in an endless loop. in the printTopology since the gateway is not available
or something else. It is not clear for us, since we have no logs nor metrics which indicate that. This commits
adds a log which is printed every 1000 failure.