feat(deps): add opensearch client and aws auth dependencies
The opensearch client dependency is used to create a client to connect to OpenSearch with. It is pretty much the exact same as the RestClient we are currently using for ElasticSearch, just managed by OpenSearch.
There is also different clients available which make the API more readable. The downside of this is that those don't provide an option to pass a JSON body to the request directly. Instead it would require a complicated mapping from the templates to a Request object. Using this client means no other changes will be required.

The awssdk auth dependency as the name suggests is used to authenticate with AWS, giving us permission to make requests to OpenSearch running in AWS and giving us access to properties used for signing the requests.