feat: make sure to de-/serialize correctly
Test de-/serialization of different manifest with different states.
Add additional checks in order to not deserialize incorrect state.
Add possibility to migrate from base interface to others, like to failed, completed etc. With additional checks.
This makes it possible to deserialize the manifest with the Interface and migrated to the expected special interface.