fix(test): fix basic authentication test
The Opensearch client does not have an implementation of the seucirty client. As a result there is no easy to use API to create users.
Instead of using a client we can make build the request ourself using the RestClient.

Note: this commit makes the test work against elastic search as we are still running the elasticsearch container. I've done this as a first step to mgirate to the opensearch clients. The container will be switched in a later commit.