fix(test): fix the RestClientFactoryTest
This test verifies that the configuration works as expected. It does this by creating the RestClient and then performing a request with a basic context. It then verifies the authentication is added to this context during the request.

With the elasticsearch RestClient this was simple to do, as it exposes the httpClient it uses via a getter. The opensearch RestClient does not do this. In order to access it I had to use reflection.