fix(test): switch to OpenSearch testcontainer
This commit switches from the ElasticSearch testcontainer to the OpenSearch testcontainer. For the majority of the tests this was straightforward. For the OpensearchClientIT it required a bit more effort.

For the OpensearchClientIT we had to start the testcontainer with the security plugin enabled. As a result SSL is enabled and we must provide a certificate to make requests to the opensearch API. Whilst we don't necessarily care for the SSL we need to enable the security in order to test basic authentication.
In order to accept all self-signed certificates we have to provide an SSL context to the RestClient as described here https://github.com/opensearch-project/opensearch-testcontainers#usage-examples. Since this nested somewhat deep in the RestClientFactory we had to add some code only meant for testing to this factory.