feat(exporter): set default AWS configuration settings
The serviceName of opensearch is defined by AWS. This is currently "es". This is not something that is likely to change, so we can hardcode this config to be "es" by default.
AWS EKS takes care of injecting some environment variables in our pod. One of these is the AWS_REGION. This is the region the pod is running in, which is also the region which is required for signing requests. We can set the region to this value by default.

Whilst we could theoretically check if the region is set to determine whether or not we are running in AWS, it's more clear that this is the case if we let users explicitly enable/disable this instead of letting this magically be the case when the AWS_REGION environment variable exists.

Note: both these settings can still be overridden if for whatever reason this is deemed necessary.