refactor(stream-platform): remote GatewayStreamer
Removes the `GatewayStreamer<M,P>` API from the stream platform.
Instead, we will move ownership of this specific API to the engine
itself, and remove the generic portion of it. This will become the
`JobStreamer` API, which will describe the behavior of the streaming
portion from the engine point of view, allowing them to define the
contract of the engine stream as far as they need it.