test(exporter): test AwsSignHttpRequestInterceptor
Adds unit tests to the interceptor that signs HTTP requests. I was hoping to make an IT, but I see no easy way to achieve this.
Instead, I opted for this unit test.

The tests verifies we can do the POST and PUT requests we do as expected. It checks whether the request has been signed with the correct data and if the content remains the same.ins as expected.