fix(backup/gcs): run gcs container with correct external url
The fake-gcs-server requires setting the external-url parameter so that
it can respond to some requests with the correct, externally accessible
url. Not setting this parameter worked fine for simple requests but
failed for more complex ones such as multipart upload where the server
response is used by the client to initiate follow-up requests.

This uses a scheme found in other testcontainers-java projects such as
Kafka and MongoDB where we inject a script to start the fake-gcs-server
with correct parameters *after* the container has already started. This
is necessary because we'll only know the external port once the
container has started. Instead of the regular entry point, we now run a
small script that waits for the real startup script to become available.