docs(exporter): rewrite Opensearch exporter README
The README was copied from the Elasticsearch exporter. It needed adjustments to make it relevant for the Opensearch exporter.

The entire structure of the configuration has been changes with this commit. Instead of having a list of bullet points I've changed it to a table giving short but clear descriptions. It also makes it very obvious what the default we use are.