refactor(backup/gcs): let fileset and manifest manager handle prefixes
Logically it makes sense that both of these components know where to
create and find their own blobs. The `GcsBackupStore` now only cares
about the basePath prefix.

This is important for the next step where we implement backup listing
because there the manifest manager needs to know the blob layout.