refactor(scheduler): add a missing methods to ConcurrencyControl
Using ConcurrentControl instead of extending from Actor is useful for writing unit tests. But some methods like "call" and "runDelayed" were missing from the interface. This commit add those missing methods.

Adding "runDelayed" to the interface resulting in some conflicts with some other classes which also extends Actor because they implemented runDelayed with a different return type. To resolve this conflict, `runDelayed` is renamed to `schedule`, as it is a more meaningful name.