feat(es-exporter): create ilm policy
The Index Lifecycle Management (ILM) Policy is needed to specify when
documents should be deleted from an index.

The policy is only created when a minimum retention age is configured.

Here is an example request:

PUT http://localhost:9200/_ilm/policy/zeebe-record-retention-policy

{
  "policy": {
    "phases": {
      "delete": {
        "min_age": "23m",
        "actions": {
          "delete": {}
        }
      }
    }
  }
}

To make sure this nested json object is created correctly, I've added
test case. Most significantly, is the required duration formatting:
https://www.elastic.co/guide/en/elasticsearch/reference/master/api-conventions.html#time-units
Note that microseconds is not a unit of java.time.Duration

If we dislike this, we could also always define it as a string.