fix: enable retention with flag
Basing whether or not to add an ILM policy to the indices on the
minimum age seems not the most user friendly. We can also just add an
enabled flag to this configuration:

```
 # new defaults
retention:
  enabled: true
  minimumAge: 30d
  policyName: zeebe-record-retention-policy
```

When I changed this, I also discovered that I could inline the
configuration pretty much everywhere, as the configuration is already
available as field. Sorry for not doing this in a separate commit, I'm
trying to make fast progress :D