ci(.github): add commitlint workflow
Adds a commitlint workflow and commitlint.config.js file to strictly
check the commit message guidelines as defined by our CONTRIBUTING.md.

We used to have GitCop control this for us, but at some point we changed
the commit types (we dropped chore), and made scopes optional. This was
not supported by GitCop, so we had to drop it.

Since then, there have been many occasions where the commit guidelines
were not considered. This is no big deal, but by applying these
consistently we maintain a easy to navigate and read git history.

The team generally believes that conventions should be automated. So
this convention should also.

I've first tried to use commitlint directly, but ended up using the most
used action for this: wagoid/commitlint-github-action. This seems
actively maintained and works well. The resulting error messages in the
workflow run logs is also clear and helpful in resolving the errors.

See: https://github.com/camunda/zeebe/blob/main/CONTRIBUTING.md/#commit-message-guidelines