refactor(transport): allow providing job specicific metrics for stream
To decouple transport from engine or job related concepts, we had previously removed job from the metric names. But this could be confusing for the users. Instead now we provide an interface for stream metrics, for which a job specific implementation can be provided by the broker. Thus the metrics names contain job to make it easier to understand.