refactor(engine): specify an error message when deployed process id not found in state
There is an issue that when a deployed process cannot be found in state, NoSuchElementException is thrown. Currently, we could not re-produce the issue. This PR adds an error message that contains persisted processes id so that we can use this id to inspect process through 'zdb' and retrieve the resource XML from there. Then, we can use that XML to reproduce the issue locally and debug it to find the actual problem.

(cherry picked from commit a816590748d9f6b02804829b14e8d6d601b0cdff)