docs(CONTRIBUTING): make commit message body optional
The body of commit messages was never mandatory, but it was documented
as such. Nobody actually checks that a body is added, and we won't
reject pull requests with commits with commits without a message body.

However, we do want to encourage contributors to write clear commit
messages, explaining both the what and why of their changes. So, I've
expanded this section of the commit message body and refer to our PR and
Code Review guidelines.